expanded bit size to 64bit

+++ cpu/memory/mapper.go
@@ -5,8 +5,8 @@ import "log"
 type Device interface {
 	Start() uint
 	End() uint
-	GetInt(addr uint16) uint16
-	SetInt(addr uint16, value uint16)
+	GetInt(addr uint64) uint64
+	SetInt(addr uint64, value uint64)
 	Remap() bool
 }
 
@@ -28,21 +28,21 @@ func (M *Mapper) find(addr uint) Device {
 	return nil
 }
 
-func (M *Mapper) GetInt(addr uint16) uint16 {
+func (M *Mapper) GetInt(addr uint64) uint64 {
 	device := M.find(uint(addr))
 	if device == nil {
 		log.Fatal("Could not find mapping for address: ", addr)
 	}
 	if device.Remap() == true {
-		addr = addr - uint16(device.Start())
+		addr = addr - uint64(device.Start())
 	}
 	return device.GetInt(addr)
 }
 
-func (M *Mapper) SetInt(addr uint16, value uint16) {
+func (M *Mapper) SetInt(addr uint64, value uint64) {
 	device := M.find(uint(addr))
 	if device.Remap() == true {
-		addr = addr - uint16(device.Start())
+		addr = addr - uint64(device.Start())
 	}
 	device.SetInt(addr, value)
 }
